code review 에 따라 code refactoring 진행
