+++ src/main/java/com/promotion/amongapi/advice/exception/AuthorizeKeyNotFoundException.java
@@ -1,4 +1,10 @@
 package com.promotion.amongapi.advice.exception;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
 public class AuthorizeKeyNotFoundException extends RuntimeException {
+    @Getter
+    String authorizeKey;
 }
