+++ src/main/java/com/promotion/amongapi/advice/exception/RequestLimitExceededException.java
@@ -1,9 +1,13 @@
 package com.promotion.amongapi.advice.exception;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class RequestLimitExceededException extends RuntimeException {
+    @Getter
     private final int requestLimit;
+    @Getter
     private final int requestCount;
 }
